test(git-hooks): group operation-state boundary coverage
