+++ test/git-hooks-pre-commit.test-support.ts
@@ -1,5 +1,12 @@
 import { execFileSync } from "node:child_process";
-import { copyFileSync, mkdirSync, symlinkSync, writeFileSync } from "node:fs";
+import {
+  copyFileSync,
+  existsSync,
+  mkdirSync,
+  readFileSync,
+  symlinkSync,
+  writeFileSync,
+} from "node:fs";
 import path from "node:path";
 import { makeTempDir as makeTempRepoRoot } from "./helpers/temp-dir.js";
 
@@ -127,3 +134,25 @@ export function stageContent(dir: string, name: string, content: string | Buffer
 }
 
 export const commitArgs = ["-c", "core.hooksPath=git-hooks", "commit", "-qm", "guard proof"];
+
+export function installFormattingRecorder(dir: string, body = ""): string {
+  const logPath = path.join(dir, "hook-tool.log");
+  writeExecutable(
+    path.join(dir, "node_modules/.bin"),
+    "oxfmt",
+    `#!/usr/bin/env bash
+set -euo pipefail
+printf 'oxfmt %s\n' "$*" >> hook-tool.log
+case "$*" in *--stdin-filepath=*) cat ;; esac
+${body}
+`,
+  );
+  return logPath;
+}
+
+export function readFormatterLog(logPath: string): string[] {
+  if (!existsSync(logPath)) {
+    return [];
+  }
+  return readFileSync(logPath, "utf8").split("\n").filter(Boolean);
+}
