+++ test/git-hooks-pre-commit.test.ts
@@ -13,7 +13,9 @@ import { afterEach, describe, expect, it } from "vitest";
 import {
   commitArgs,
   createContentGuardFixture,
+  installFormattingRecorder,
   installPreCommitFixture,
+  readFormatterLog,
   literals,
   rulePath,
   ruleSetting,
@@ -25,21 +27,6 @@ import {
 import { cleanupTempDirs, makeTempDir as makeTempRepoRoot } from "./helpers/temp-dir.js";
 
 const tempDirs: string[] = [];
-
-function installFormattingRecorder(dir: string, body = ""): string {
-  const logPath = path.join(dir, "hook-tool.log");
-  writeExecutable(
-    path.join(dir, "node_modules/.bin"),
-    "oxfmt",
-    `#!/usr/bin/env bash
-set -euo pipefail
-printf 'oxfmt %s\n' "$*" >> hook-tool.log
-case "$*" in *--stdin-filepath=*) cat ;; esac
-${body}
-`,
-  );
-  return logPath;
-}
 
 function installRunNodeToolFixture(dir: string): void {
   mkdirSync(path.join(dir, "scripts", "pre-commit"), { recursive: true });
@@ -57,37 +44,6 @@ function splitNonEmptyLines(output: string): string[] {
     }
   }
   return lines;
-}
-
-function readFormatterLog(logPath: string): string[] {
-  if (!existsSync(logPath)) {
-    return [];
-  }
-  return splitNonEmptyLines(readFileSync(logPath, "utf8"));
-}
-
-function createOperationFixture(linked: boolean, revert = false): { dir: string; primary: string } {
-  const primary = createContentGuardFixture(tempDirs);
-  const commit = (name: string, content: string) => {
-    stage(primary, name, content);
-    run(primary, "git", ["commit", "-qm", "operation fixture"]);
-  };
-  commit("changed.ts", "export const value = 0;\n");
-  run(primary, "git", ["checkout", "-qb", "side"]);
-  commit("changed.ts", "export const value = 1;\n");
-  commit("next.txt", "next step\n");
-  run(primary, "git", ["checkout", "-q", "main"]);
-  if (revert) {
-    run(primary, "git", ["merge", "--ff-only", "side"]);
-  }
-  commit("changed.ts", "export const value = 2;\n");
-  let dir = primary;
-  if (linked) {
-    dir = path.join(primary, "linked");
-    run(primary, "git", ["worktree", "add", "-qb", "linked", dir, "HEAD"]);
-    installPreCommitFixture(dir);
-  }
-  return { dir, primary };
 }
 
 afterEach(() => {
@@ -115,119 +71,6 @@ describe("git-hooks/pre-commit (integration)", () => {
 
     const staged = splitNonEmptyLines(run(dir, "git", ["diff", "--cached", "--name-only"]));
     expect(staged).toEqual(["--all"]);
-  });
-
-  describe.each([false, true])("Git operation state (linked=%s)", (linked) => {
-    it.each([
-      { operation: "merge", args: ["merge", "--no-commit", "side"], state: "MERGE_HEAD" },
-      {
-        operation: "rebase merge",
-        args: ["rebase", "--merge", "side"],
-        state: "rebase-merge/git-rebase-todo",
-      },
-      {
-        operation: "rebase apply",
-        args: ["rebase", "--apply", "side"],
-        state: "rebase-apply/next",
-      },
-      { operation: "cherry-pick", args: ["cherry-pick", "side~1"], state: "CHERRY_PICK_HEAD" },
-      { operation: "revert", args: ["revert", "--no-edit", "side~1"], state: "REVERT_HEAD" },
-      {
-        operation: "cherry-pick sequencer-only",
-        args: ["cherry-pick", "side~1", "side"],
-        state: "sequencer/todo",
-      },
-      {
-        operation: "revert sequencer-only",
-        args: ["revert", "--no-edit", "side~1", "side"],
-        state: "sequencer/todo",
-      },
-    ])("preserves operation staging during $operation", ({ operation, args, state }) => {
-      const { dir, primary } = createOperationFixture(linked, operation.startsWith("revert"));
-      expect(runFailure(dir, "git", args).status).toBe(1);
-      expect(run(dir, "git", ["ls-files", "--unmerged"])).not.toBe("");
-      const metadata = path.resolve(dir, run(dir, "git", ["rev-parse", "--git-path", state]));
-      if (linked) {
-        expect(existsSync(path.join(primary, ".git", state))).toBe(false);
-      }
-      if (state === "sequencer/todo") {
-        stage(dir, "changed.ts", "export const value = 3;\n");
-        // A real resolution commit clears the per-step ref while later steps remain queued.
-        run(dir, "git", commitArgs);
-        for (const ref of ["CHERRY_PICK_HEAD", "REVERT_HEAD", "REBASE_HEAD"]) {
-          expect(
-            existsSync(path.resolve(dir, run(dir, "git", ["rev-parse", "--git-path", ref]))),
-          ).toBe(false);
-        }
-      }
-      const before = readFileSync(metadata);
-      const staged = "export const value=4;\n";
-      const unstaged = `${staged}// unstaged edit\n`;
-      stage(dir, "changed.ts", staged);
-      const stagedOid = run(dir, "git", ["rev-parse", ":changed.ts"]);
-      writeFileSync(path.join(dir, "changed.ts"), unstaged);
-      const log = installFormattingRecorder(
-        dir,
-        "printf '// formatted\\n' >> changed.ts\nprintf formatted",
-      );
-
-      expect(run(dir, "bash", ["git-hooks/pre-commit"])).toBe("");
-      expect(readFormatterLog(log)).toEqual([]);
-      expect(run(dir, "git", ["rev-parse", ":changed.ts"])).toBe(stagedOid);
-      expect(readFileSync(path.join(dir, "changed.ts"), "utf8")).toBe(unstaged);
-      expect(readFileSync(metadata)).toEqual(before);
-
-      stage(dir, "changed.ts", literals[1]);
-      expect(runFailure(dir, "bash", ["git-hooks/pre-commit"]).stderr).toContain(
-        "Blocked staged content",
-      );
-      stage(dir, "changed.ts", staged);
-      writeFileSync(path.join(dir, "changed.ts"), unstaged);
-      run(dir, "git", commitArgs);
-      expect(readFormatterLog(log)).toEqual([]);
-      expect(run(dir, "git", ["rev-parse", "HEAD:changed.ts"])).toBe(stagedOid);
-      expect(readFileSync(path.join(dir, "changed.ts"), "utf8")).toBe(unstaged);
-    });
-
-    it("formats and restages with an orphan REBASE_HEAD", () => {
-      const { dir, primary } = createOperationFixture(linked);
-      if (linked) {
-        // Another worktree's real rebase must not suppress this worktree's ordinary commit.
-        expect(runFailure(primary, "git", ["rebase", "--merge", "side"]).status).toBe(1);
-      }
-      const head = run(dir, "git", ["rev-parse", "HEAD"]);
-      // Reproduce the observed orphan state without claiming how it was left behind.
-      run(dir, "git", ["update-ref", "REBASE_HEAD", head]);
-      for (const state of [
-        "MERGE_HEAD",
-        "CHERRY_PICK_HEAD",
-        "REVERT_HEAD",
-        "rebase-merge",
-        "rebase-apply",
-        "sequencer",
-      ]) {
-        expect(
-          existsSync(path.resolve(dir, run(dir, "git", ["rev-parse", "--git-path", state]))),
-        ).toBe(false);
-      }
-      expect(runFailure(dir, "git", ["rebase", "--continue"]).stderr).toContain(
-        "no rebase in progress",
-      );
-      const working = "export const value=5;\n";
-      stage(dir, "changed.ts", working);
-      const log = installFormattingRecorder(
-        dir,
-        "printf '// formatted\\n' >> changed.ts\nprintf formatted",
-      );
-      expect(run(dir, "bash", ["git-hooks/pre-commit"])).toBe("formatted");
-      expect(readFormatterLog(log)).toEqual([
-        "oxfmt --write --no-error-on-unmatched-pattern changed.ts",
-      ]);
-      expect(readFileSync(path.join(dir, "changed.ts"), "utf8")).toBe(`${working}// formatted\n`);
-      expect(run(dir, "git", ["show", ":changed.ts"])).toBe(`${working}// formatted`);
-      expect(run(dir, "git", ["diff", "--", "changed.ts"])).toBe("");
-      expect(run(dir, "git", ["rev-parse", "REBASE_HEAD"])).toBe(head);
-    });
   });
 
   it.each(["configured", "unconfigured", "external"])(
